convoydriver: Add opts parameter to Umount() and Mountpoint()
To make it consistent with Docker go-plugins-helpers
